Refactor existing DynamicFieldMapperTests to an abstract class and implementation
