+++ server/src/test/java/org/elasticsearch/index/mapper/NonDynamicFieldMapperTests.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+package org.elasticsearch.index.mapper;
+
+import org.elasticsearch.plugins.MapperPlugin;
+import org.elasticsearch.plugins.Plugin;
+
+import java.io.IOException;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+
+public class NonDynamicFieldMapperTests extends NonDynamicFieldMapperTestCase {
+    @Override
+    protected Collection<Class<? extends Plugin>> getPlugins() {
+        return List.of(NonDynamicFieldPlugin.class);
+    }
+
+    protected String getTypeName() {
+        return NonDynamicFieldMapper.NAME;
+    }
+
+    protected String getMapping() {
+        return """
+           "type": "%s"
+           """.formatted(NonDynamicFieldMapper.NAME);
+    }
+
+    public static class NonDynamicFieldPlugin extends Plugin implements MapperPlugin {
+        public NonDynamicFieldPlugin() {}
+
+        @Override
+        public Map<String, Mapper.TypeParser> getMappers() {
+            return Map.of(NonDynamicFieldMapper.NAME, NonDynamicFieldMapper.PARSER);
+        }
+    }
+
+    private static class NonDynamicFieldMapper extends FieldMapper {
+        private static final String NAME = "non_dynamic";
+
+        private static final TypeParser PARSER = new TypeParser(
+            (n, c) -> new Builder(n),
+            List.of(notFromDynamicTemplates(NAME), notInMultiFields(NAME))
+        );
+
+        private static class Builder extends FieldMapper.Builder {
+            private final Parameter<Map<String, String>> meta = Parameter.metaParam();
+
+            Builder(String name) {
+                super(name);
+            }
+
+            @Override
+            protected Parameter<?>[] getParameters() {
+                return new Parameter<?>[] { meta };
+            }
+
+            @Override
+            public NonDynamicFieldMapper build(MapperBuilderContext context) {
+                return new NonDynamicFieldMapper(name(), new TextFieldMapper.TextFieldType(name(), false, true, meta.getValue()));
+            }
+        }
+
+        private NonDynamicFieldMapper(String simpleName, MappedFieldType mappedFieldType) {
+            super(simpleName, mappedFieldType, MultiFields.empty(), CopyTo.empty());
+        }
+
+        @Override
+        protected String contentType() {
+            return NAME;
+        }
+
+        @Override
+        protected void parseCreateField(DocumentParserContext context) throws IOException {}
+
+        @Override
+        public FieldMapper.Builder getMergeBuilder() {
+            return new Builder(simpleName()).init(this);
+        }
+    }
+}
